Add everything
This first commit contains what I think is necessary and definitely some stuff that is not. For example, in the scripts section there is good tools like the NCPython downloader for the GLM data, and the MultiProcessGLMData script that processes it, but the BetterMap might not be necessary.
One main problem is that some of these tools/notebooks are greatly documented and some of them are not. I haven't had the most time in the world to document everything, and some things are going to be better than others, and that will be updated/removed when I can.